+++ packages/wallet/database/migrations/create_wallet_recharges_table.php
@@ -33,7 +33,8 @@ return new class extends Migration {
             $table->unsignedBigInteger('payment_id')->nullable()->comment('支付单ID');
             $table->string('payment_channel_trade_no', '64')->nullable()->comment('支付渠道单号');
             $table->string('payment_mode', '32')->nullable()->comment('支付方式');
-
+            $table->string('fail_reason')->nullable()->comment('失败原因');
+            $table->string('fail_code', 32)->nullable()->comment('失败代码');
             $table->json('extra')->nullable()->comment('扩展字段');
             $table->operator();
             $table->comment('钱包-充值单');