+++ packages/wallet/src/Application/Services/Recharge/Commands/CreateRechargeCommand.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace RedJasmine\Wallet\Application\Services\Recharge\Commands;
+
+use Cknow\Money\Money;
+use RedJasmine\Support\Contracts\UserInterface;
+use RedJasmine\Wallet\Domain\Data\WalletRechargeData;
+
+class CreateRechargeCommand extends WalletRechargeData
+{
+
+
+
+}