+++ packages/wallet/src/UI/Http/User/Api/Controllers/WalletRechargeController.php
@@ -6,11 +6,14 @@ use Illuminate\Http\Request;
 use RedJasmine\Support\Http\Controllers\UserOwnerTools;
 use RedJasmine\Support\UI\Http\Controllers\RestControllerActions;
 use RedJasmine\Support\UI\Http\Controllers\RestQueryControllerActions;
-use RedJasmine\Wallet\Application\Commands\Recharge\CreateRechargeCommand;
+
 use RedJasmine\Wallet\Application\Queries\Recharge\WalletRechargeDetailQuery;
 use RedJasmine\Wallet\Application\Queries\Recharge\WalletRechargeListQuery;
+use RedJasmine\Wallet\Application\Services\Recharge\Commands\CreateRechargeCommand;
 use RedJasmine\Wallet\Application\Services\Recharge\Queries\WalletRechargePaginateQuery;
 use RedJasmine\Wallet\Application\Services\Recharge\WalletRechargeApplicationService;
+use RedJasmine\Wallet\Application\Services\Wallet\Queries\FindByOwnerTypeQuery;
+use RedJasmine\Wallet\Application\Services\Wallet\WalletApplicationService;
 use RedJasmine\Wallet\Domain\Models\WalletRecharge;
 use RedJasmine\Wallet\UI\Http\User\Api\Requests\CreateRechargeRequest;
 use RedJasmine\Wallet\UI\Http\User\Api\Resources\WalletRechargeResource;
@@ -26,6 +29,7 @@ class WalletRechargeController extends Controller
 
     public function __construct(
         protected WalletRechargeApplicationService $service,
+        protected WalletApplicationService $walletApplicationService,
     ) {
         // 设置查询作用域
         $this->service->readRepository->withQuery(function ($query) {
@@ -39,13 +43,18 @@ class WalletRechargeController extends Controller
      */
     public function store(CreateRechargeRequest $request)
     {
-        $command                = new CreateRechargeCommand();
-        $command->owner         = $this->getOwner();
-        $command->walletType    = $request->get('wallet_type');
-        $command->amount        = $request->get('amount');
-        $command->paymentMethod = $request->get('payment_method');
-        $command->remark        = $request->get('remark');
-        $recharge               = $this->service->create($command);
+
+
+        $query        = new FindByOwnerTypeQuery();
+        $query->owner = $this->getOwner();
+        $query->type  = $request->wallet_type;
+
+        $wallet = $this->walletApplicationService->findByOwnerType($query);
+
+        $command = CreateRechargeCommand::from($request);
+
+        $command->setKey($wallet->id);
+        $recharge = $this->service->create($command);
 
         return new WalletRechargeResource($recharge);
     }